+++ lib/opal/cli_runners.rb
@@ -20,7 +20,6 @@ module Opal
 end
 
 require 'opal/cli_runners/applescript'
-require 'opal/cli_runners/phantomjs'
 require 'opal/cli_runners/nodejs'
 require 'opal/cli_runners/server'
 require 'opal/cli_runners/nashorn'