+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Whitespace conventions:
 - Added `Opal::SimpleServer` as the quickest way to get up and running with Opal: `rackup -ropal -ropal/simple_server -b 'Opal.append_path("app"); run Opal::SimpleServer.new'`
 - Added `String#ascii_only?` (#1592)
 - Added `StringScanner#matched_size` (#1595)
+- Added `Hash#compare_by_identity` (#1657)
 
 
 ### Removed