+++ CHANGELOG.md
@@ -21,6 +21,16 @@ Whitespace conventions:
 
 
 
+## [0.10.4] - 2017-06-21
+
+
+### Fixed
+
+- Fix `Time#zone` for zones expressed numerically
+
+
+
+
 ## [0.10.3] - 2016-09-09
 
 
@@ -881,6 +891,7 @@ Whitespace conventions:
 
 
 
+[0.10.4]: https://github.com/opal/opal/compare/v0.10.3...v0.10.4
 [0.10.3]: https://github.com/opal/opal/compare/v0.10.2...v0.10.3
 [0.10.2]: https://github.com/opal/opal/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/opal/opal/compare/v0.10.0...v0.10.1